update code to set indoor nodes
This update enables to create an indoor area where nodes belong to an indoor environment. Every time a new node's location is generated, it is checked to be inside the indoor area and indoor parameter is set. Origin and size of the indoor area are fixed, but they can easily be changed. 
Indoor Area will be defined only if indoor_nodes is True in GlobalConfig.py.